Create file for email authentication data

+++ install/index.php
@@ -2,8 +2,7 @@
 /**
  * Author:    Aleksandar Vranešević
  * URI:       https://vavok.net
- * Updated:   01.08.2020. 19:11:08
-*/
+ */
 
 if (isset($_GET['step'])) {
 	$step = $_GET['step'];
@@ -22,22 +21,22 @@ if (empty($step)) {
     <p>By installing this software you agree to <a href="../LICENSE">license</a><br /><br /></p>
 
     <form method="post" action="index.php?step=first_end">
-    <fieldset>
-    <legend>Installation: first step</legend>
-    <label for="dbhost">Database host (mostly localhost):</label><br />
-    <input name="dbhost" id="dbhost" maxlength="30" /><br />
-    <label for="database">Database name:</label><br />
-    <input name="database" id="database" maxlength="30" /><br />
-    <label for="dbusername">Database username:</label><br />
-    <input name="dbusername" id="dbusername" maxlength="30" /><br />
-    <label for="dbpass">Database password:</label><br />
-    <input name="dbpass" id="dbpass" maxlength="30" /><br />
-    <label for="select">Language:</label>
-    <select name="language" id="language">
-    <option name="english" value="english">English</option>
-    </select>
-    <input value="Continue..." type="submit" />
-    </fieldset>
+        <fieldset>
+            <legend>Installation: first step</legend>
+            <label for="dbhost">Database host (mostly localhost):</label><br />
+            <input name="dbhost" id="dbhost" maxlength="30" /><br />
+            <label for="database">Database name:</label><br />
+            <input name="database" id="database" maxlength="30" /><br />
+            <label for="dbusername">Database username:</label><br />
+            <input name="dbusername" id="dbusername" maxlength="30" /><br />
+            <label for="dbpass">Database password:</label><br />
+            <input name="dbpass" id="dbpass" maxlength="30" /><br />
+            <label for="select">Language:</label>
+            <select name="language" id="language">
+            <option name="english" value="english">English</option>
+            </select>
+            <input value="Continue..." type="submit" />
+        </fieldset>
     </form>
     <hr />
 
@@ -59,17 +58,44 @@ if ($step == 'first_end') {
     /**
      * Create configuration .env file
      */
-$config_data = 'DB_HOST=localhost
+$config_data = <<<EOD
+DB_HOST=localhost
 DB_USERNAME=root
 DB_PASSWORD=root
 DB_DATABASE=database
 DB_PREFIX=null
 
-SITE_STAGE=null';
+SITE_STAGE=null
+EOD;
 
-    if (!file_exists('../.env')) {
-        file_put_contents('../.env', $config_data);
-    }
+    if (!file_exists('../.env')) file_put_contents('../.env', $config_data);
+
+    /**
+     * Create email authentication file .available_emails.php
+     */
+$available_emails = <<<'EOT'
+<?php
+/**
+ * Mail accounts that have authentication data
+ */
+$available_mails = array(
+    array(
+        'username' => 'example_email@example.com',
+        'password' => 'type_password_here',
+        'port' => 587,
+        'host' => 'mail.example.com'
+    ),
+    array(
+        'username' => 'another_email@example.com',
+        'password' => 'type_password_here',
+        'port' => 587,
+        'host' => 'mail.example.com'
+    )
+);
+?>
+EOT;
+
+    if (!file_exists('../include/.available_emails.php')) file_put_contents('../include/.available_emails.php', $available_emails);
 
     include "../include/classes/Config.class.php";
 
@@ -99,7 +125,7 @@ SITE_STAGE=null';
     echo '<p><img src="../images/img/partners.gif" alt="" /> Second step<br></p>';
 
     echo '<p><img src="../images/img/reload.gif" alt="" /> Data successfully saved!</p>';
-    echo '<p><a href="finish.php?step=second&amp;host=' . $mysql_host . '&amp;user=' . $mysql_username . '&amp;pass=' . $mysql_password . '&amp;db=' . $mysql_database . '">Second step</a> - Creating database</p>';
+    echo '<p><a href="finish.php?step=second&host=' . $mysql_host . '&user=' . $mysql_username . '&pass=' . $mysql_password . '&db=' . $mysql_database . '">Second step</a> - Creating database</p>';
 
 }
